requirements.txt
pandas==1.3.4
python-dotenv==0.19.2
dash==2.1.0
dash-bootstrap-components==1.0.2